+++ src/components/_home/green-magazine/GreenMagazineCarousel.tsx
