+++ src/components/_home/green-magazine/GreenMagazineSection.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { MagazinePostData } from '@/app/green-magazine/_types/magazine.types';
-import { GreenMagazineCarousel } from '@/components/_home/GreenMagazineCarousel';
+import { GreenMagazineCarousel } from '@/components/_home/green-magazine/GreenMagazineCarousel';
 import { magazineForHome } from '@/lib/functions/home/magazineForHome';
 import { useEffect, useState } from 'react';
